requirements.txt
django
fastapi
uvicorn[standard]
mysqlclient
python-dotenv
gunicorn
pymysql
